pyproject.toml
[tool.poetry]
name = "python w zastosowaniach naukowych"
version = "0.1.0"
description = ""
authors = ["Daniel Krajewski"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.9.2"
tqdm = "^4.66.5"
ascii-graph = "^1.5.1"
numpy = ">=1.24,<2.2"
argparse = "^1.4.0"
pillow = "^11.1.0"
numba = "^0.61.0"
matplotlib = "^3.10.0"
jupyterlab = "^4.3.4"
beautifulsoup4 = "^4.13.3"
bs4 = "^0.0.2"
selenium = "^4.28.1"
scipy = "^1.15.1"
sympy = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
